Sign-in with google functionality added

+++ php/delete.php
@@ -4,7 +4,9 @@ $username = "root";
 $password = "";
 $dbname = "db_expensetracker";
 
-$item = file_get_contents('php://input');
+//$item = file_get_contents('php://input');
+$item = $_POST['del'];
+$loginid = $_POST['loginid'];
 
 // Create connection
 $conn = new mysqli($servername, $username, $password, $dbname);
@@ -14,7 +16,7 @@ if ($conn->connect_error) {
 }
 
 // sql to delete a record
-$sql = "DELETE FROM tbl_expenses WHERE item='$item'";
+$sql = "DELETE FROM tbl_expenses WHERE item='$item' and UserID='$loginid'";
 
 if ($conn->query($sql) === TRUE) {
     echo "Record deleted successfully";
